# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: OmniLock Backend API
  version: 0.1.0
  description: Software-side API for OmniLock smart locker backend (Node.js + Express + PostgreSQL).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceSecret:
      type: apiKey
      in: header
      name: x-device-secret
    webhookSignature:
      type: apiKey
      in: header
      name: x-webhook-signature
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200': { description: User profile }
        '401': { description: Unauthorized }
  /me:
    get:
      tags: [Auth]
      summary: Alternate current user profile endpoint
      security:
        - bearerAuth: []
      responses:
        '200': { description: User profile }
...